Move deployed_dashboard_url into common.dashboard
